gmseq v0.04
I identified 2 bugs. The first, which explains the missing values for n=7,
was because I had prematurely optimised append_divisors() in the belief that
it would only need divisors < 2^32. In fact it needs divisors up to 3.2^32
for n=7 (and up to 3.2^64 for n=8).

I have fixed this by writing two versions of the relevant routines, one to
use unsigned longs (called whenever possible), and the other to use mpz_t.

The second was an error in the PARI code, in the one place where we used
floating point arithmetic - the calculation of cmax. I think the first
example where it failed is in:
  floor(1/(sqrtn(21^3/20^3, 3) - 1))
which returned 19 instead of 20. Because of that the PARI code failed to
spot the solution { 7, 7, 9, 9, 14, 20, 20, 20 } for n=8. The GMP code
does not have this problem: it is using a different approach that sticks
to integer calculations throughout.