gmseq v0.06
Gained some major savings from a) testing if d fits in a ulong, and taking
advantage of that in count_mod and count_min_mod when testing for f == -q
(mod d) and b) small tweaks to append_udivisors_pp. Also gained smaller
amounts by implementing SMALLMOD, to avoid the full list of divisors in
count_mod when d is small enough.

Even with those changes, the bulk of the time for n=8 cases is still in
count_mod() and count_mod_min(), and in their calls (via append_udivisors)
to append_udivisors_pp. This could be made far faster if there was a way
to discover factors of pq == -q (mod p-q) more easily than by constructing
the full list and then testing each value.

Unexpectedly, increasing the defined value of SMALLMOD doesn't improve
things further: raising it from 32 to 49, it catches an extra 6% of the
cases that would otherwise use LAT ("list and test"), but it reduces the
time taken in LAT code only by 1.8%. That seems to imply that LAT is on
average slower for some reason for the cases when d is larger, so that
removing some cases when d is small has a smaller than expected effect -
I'm not sure why that should be.

count_min_mod() cannot use the SMALLMOD trick - we need the magnitude
of each factor so we can check against minf.