Rewrite A001055/A001055 using composition of partitions
For the list of powers [x, y, z], we take the previously cached data for
the list [x, y] and compose them with the partitions of z.

This is rather more complex code, but way faster: our previous limit at
n=217728000 which took about 17250 CPU-seconds now takes about 10s.
We can now reach 6e10 in 68s, at which point memory usage is 109MB.