Export solution handling
Make a non-inlined function to record solutions and update the pp list
for the next length; call that function inline rather than breaking out
of the index loop. This now simplifies the index loop (we now only ever
'continue', and break out only when we exhausted the solution space) and
removes the need for the outer lengthening loop.

Also removed best_offset, which was never used, and gets our struct size
to 8 words.