Shortcircuit all found=0, found=1
Separate out the number of new hits found for a given prime; this gives
a microoptimization (since we avoid adjusting excess and needed by
independent increments), and also lets us shortcircuit whenever found=0
(since we can always do better), and collapse all found=1 for a given
prime to a single call (since all such are equivalent).

This gives a speedup in excess of 10x over the previous version, which
collapsed only a proportion of the found=1 cases.