Correct and simplify results output
In results, primes subject to the singleton optimization were shown as
if they were set to either the prime itself (which was at least obvious)
or to the first offset at which they matched a single value (which was
wrong). Correct that, and show the offsets in a streamlined form that
better highlights which are floating.

Timings not recalculated, this variant requires about one extra store
operation per iteration.