divrep FIXME added
Trying to find a limit on k for a particular n by binary chop is
made a lot slower by the cascading propagation of denied values
to other moduli, getting worse at least quadratically as the -c<x>
option is increased.

This can be fixed, at least in part, by checking for it earlier;
it is likely also this would allow a better diagnostic (since
currently it almost always reports either "all values (mod 2)
disallowed" or "all values (mod 1)").