dpart: combine search and insert in avl_tree
Since the only interface we need is avl_seen(), it seems wasteful to search
the tree once for the avl_exists() check, and then again for the avl_insert().
Currently the proportion of inserts is very small, so the saving here is not
significant; but as we improve pruning, the proportion should grow.