dpart: record the distinct transformations of each piece
Record a list of the symmetry maps the yield distinct variants of each
piece; use this to reduce the expensive duplication of effort while
iterating over pieces.

Since the data can get large, we do extra work up front to ensure that
we also de-dup the lists this generates: at n=5, we'd otherwise need
about 10GB to store the lists.