dpart: prune duplicates
Now that we avoid filling in the size-1 pieces, the set we check looks
the same as the partial set at that stage of the recursion; therefore,
if on checking it as a solution we find we've seen it before, we can
immediately prune with no more effort.

So make check_solution() return a boolean to indicate if this is a new
solution, and move the "fill with size-1" solution to before rather than
after the loop in which we try to insert other size pieces.

This improves n=4 from 8.9s to 3.3s.