dpart: try one I-partition at a time
Try to shard the solution space more effectively by first picking
an integer ("I-") partition of NODES, then finding only partitions
of the shape ("S-partition) into pieces of the sizes given by the
I-partition.

We are able to get the correct results still, but the new code takes
much longer to do the same work, and does not give the hoped-for saving
of memory. (For pieces up to size 2 for n=5, this took 7.5h instead of 1h
to find the 1+3635430 solutions.)