dpart: replace seenness with canonicalness for pruning (*BROKEN*)
When recursing, skip if the current subset is not canonical on the
assumption that the subset will appear canonically either earlier or
later. This does away with the complexity and memory cost of storing
what has been seen.

Currently this misses some solutions for n>=4, because the canonicalization
is not progressive: there are canonical C<k+1>-piece partials that have no
canonical C<k>-piece predecessor. Fixing that will require a more complex
canonicalization (which will probably be slower). We are fast right now
though: we find 68,524 solutions (of 75,218) for n=4 in 1.1s.

Memory is also no longer a problem now: we can spend 30ks finding the first
350 million solutions for n=5, and the process grows only to 2.6MB.