Recover missing files: pieces.[ch], sym_set.[ch]
Not sure what happened here, but git lost all trace of these files that
had been developed over the last 50-or-so commits.

This is the version relevant to the about-to-be-committed support for
separating generation of piece and sym-set data from its use in part<n>;
will try to work out later if it's possible to recover or recreate the
prior history.