Separate creation of piece/sym_set data from usage
New programs build<n> will generate the information on pieces, and write
to a file; existing part<n> programs will read from the generated file.

This is an intermediate step; next stage is to annotate the piece data
with counts of the number of ways each can be partitioned, and to take
advantage of that in partition calculation.