A101877: v4: cache next and least found matches
Next step: find() redoes an awful lot of work finding each successive
match. Given the require modulus $mod and the number of elements to
discard $count, we now save the array of indices @nfm leading to the
*next found match*, and $lfm, the sum of the *least found match* greater
than the last one actually found.

Then, to find the next match:
- find $min = min(grep defined, {$mod}{$count}->{lfm})
- for each lfm not defined:
  - use the existing multiloop algorithm to find the next lfm
  - aborting if we cannot beat $min
  - but seed it with @nfm as the starting indices
  - and return @nfm as the starting indices of the first found match
    (if different)
  - and replace $min if $lfm is lower
- then delete the selected {$mod}{$count}->{lfm} and return it
- not forgetting to watch out for duplicate sums

Now finds a(5)=184 in 3.3s and a(6)=469 in 117.7s, and able to narrow
down the possibilities for a(7) to the range 1220-1251.