divrep/harness: know about primes
TauG entries now record whether their 'n' is prime, and use this to
set a more refined base priority: we're less interested in chasing
primes beyond the first few, since beyond the first dozen or so results
they're very hard to discover.

Base priority is now:
  - log_2(n)
  - 10 floor(log_10(n))
  - 10 c(n > 23 and is_prime(n))

We'll also take advantage of the information when picking strategy
in a later commit.