divrep/harness: spot if a run logged a fix_power()
fix_power is a very powerful optimization, so spot and record when it
has been used.

When it is in use, prep time grows much more quickly as the -c option
is raised; however if we reduce it too far we may fail to spot the
fix_power opportunity, and run time balloons.

A subsequent commit will take advantage of this to tweak strategy.