divrep/harness: more strategy
Refine calculation of the -c option, since it is critical for efficiency.

For "simple" runs (limit < 2 ^ 20), we use -c100.

For other cases, we look at the previous run as a base point.

For "fast" runs (where expected prep + run time is less than 10s, based
on previous run), we set the -c option to half what it was on the previous
run, subject to a minimum of 100; for others, we set it to
100 + 50 * log_2(limit), dividing that by 4 if the previous run used
a fix_power.

If:
  the resulting -c value is more than the previous one, and the prep time
  exceeds the expected run time
or:
  the resulting -c value is less than the previous one, and the prep time
  is less than a tenth of the expected run time
then:
  move the -c value halfway to the previous value.

Finally, raise it to at least the recorded "minimum c" value for this
TauF record (currently only available by setting manually).

This probably sounds chaotic, but it's an attempt at some heuristics
that don't involve huge amounts of calculation or additional runs
while avoiding some observed pathological cases. A subsequent step
will be to do sequences of optimization-only runs, but that will
probably lead to yet more complexity.