gtauseq: "test all" option
If the new -ta option is specified, we check each of the testers on
each of the candidate numbers, and report for each how many of the
candidates it managed to reject.

Using this to determine the -ts order for subsequent runs should
be a useful optimization.

The information appears in the 301 output lines: if such a line
ends with 'seen [1000 726 501 988]' that implies that 1000 candidate
numbers were checked, and testers 1, 2 and 3 respectively rejected
726, 501, 988 of the candidates. As such subsequent runs should
use -ts3,1,2.