divrep/harness: optimize test order
Store the desired test order in tauf; if the expected binary runtime
exceeds USE_TS (20 minutes), request optimization for the next run.

For an optimizing run, we supply the -ta flag to gtauseq, parse
the resulting order and store it in tauf. Whenever a subsequent run
sees we have a known test order, we supply it as the -ts option.