squareloop: initial code
Find n such that there exists a loop of the integers 1 .. n in which each
pair of consecutive values sums to a square.

Bonus: find also n where a broken loop exists.

This code finds first broken loop at n=17, with the sequence:
  16 9 7 2 14 11 5 4 12 13 3 6 10 15 1 8 17

Since the first squares > 18 are 25, 36, 49, this implies the first loop
must have n >= 31.