gtauseq: better fix_power detection
When p divides both k and n, we may have powers that are fixed that we
were not detecting: when n+kd is known to have a factor p^x, then if
p^{x+1} divides n and 0 (mod p) is disallowed, it is fixed.

This improves fix_power detection for example with (637, 14) or with
(575, 6) when -m5!0.