A337663: implement coalesce
This allows two groups to be joined when necessary, with a common location
in which the next value will be placed, and optionally additional 1s in
its neighbourhood.

This should in principle allow calculation up to a(8); at a(9) we may
need to be able to coalesce 3 groups at the same time.

I don't think I can actually run this to find a(5) though, will probably
need to reimplement in C now I have a reference implementation.