A337663: more symmetry
Regularize interface for sym_transloc, to take (x, y) rather than a group;
add sym_checkloc(), true if the supplied loc is invariant under the
symmetry; update callers.

In coalesce, check if the location on the second group is invariant
under some symmetry of that group; where it is, skip transforming for
that symmetry. (Saves about 0.5% in the first 2e6 boards.)

What about the converse, when the location on the first group is invariant
under some symmetry of the first group, which transforms of the second
group can we skip in that case?