A337663: Reduce more symmetries in coalesce
In coalesce_group we check for possible matches of the first group against
each orientation of the second group. If the second group has symmetries,
we already winnow the transformations we try to avoid duplication.

This change now checks symmetries of the first group, and avoids
orientations of the second group that would give duplicate results
when combined.