A337663: allow run to be resumed
Every 100th board (at least when there are still unused - when there are
none, it is assumed there is little work to do) we write a line showing
the current best, the full history of how we got to this board, and details
of the board itself. The current best and history can be used to resume
a run.

The command line arguments are now:
  ./cA337663 <n> <freq> <history>
.. where 'freq' is how many boards (with unused) to traverse before
printing - it defaults to 100, but on my computer 1000000 seems a more
useful count;
.. where 'history' is a string (usually derived from the output of an
earlier run) which starts with the current best and history. The code
assumes the first entry starting with a non-digit terminates the history.

Additionally, we now show each board that improves ("best") or equals
("ebest") the previous best.