divrep: more careful initial value calculation under fixpower
We attempted to find an initial value for the power-root y based on the
requested starting value d by calculating y = 1 + floor(_dtoy(d - 1));
however when the multiplier x and divisor k in the relation d=(xy^z-n)/k
could have a common factor, this could result in a value for which we
could not perform the reverse _ytod calculation.

Modifying it instead to find y = g + floor(_dtoy(d - g)), where g = gcd(x,k)
should fix the problem.