divrep: improve handling of large fixed moduli
When large moduli are fixed, usually with a -m or similar option,
avoid creating useless large bitvectors, just make sure the information
is incorporated in mult/mod_mult and in the bitvectors in the range
we're actually going to look at.

This creates a slight problem if the fix is deferred; in that case we
discard the parts we can't reach, with a warning.