harness: don't delete impossible tauf entries
A given entry in tauf may already have linked records in run at the point
we discover f(n,k) is impossible, and we don't want to delete those
records. So at the point we discover it is impossible, simply mark it
as such rather than trying to delete it.