A051602: initial commit, somewhat buggy
Given a selection of n points on a plane, what arrangements give
the greatest number of squares (with a point at each corner)?

We assume all such maximal arrangements can be generated by starting
with a unit square and iteratively extending with a new square by
adding either 1 or 2 points.

We are interested in the maximum number of squares, and also the
minimum and maximum grid sizes yielding maximal arrangements.

Known issues:
- C code miscalculates grid sizes
- C and Perl code miss some arrangements
  - we must try 1-point extensions first, and can suppress 2-point
    extensions where either point has been tried as a 1-point (not v.v.)
  - we cannot jerry-rig symmetry suppression by starting try2 at [3, 2, 0],
    this misses arrangements in which more than one of the starting pairs
    is to be extended
- C code runs out of memory on n=17, we have to use list-walking to search
  instead of trying to create a full x by y lookup. (Irrelevant for perl
  code, it's too slow to reach n=17.)