A051602: restore the missing arrangements
In an attempt to jerry-rig some symmetry deduplication, we start the
search for pairs to extend on the fourth edge of the initial square.
However that causes us to miss arrangements in which we have to
extend more than one of those edges, so we'll have to recover the
deduplication some other way.