A051602: oops, fix is_transpose()
Hidden constant needed to be changed when switching from an enum to
bit values for the symmetry types.

So instead of (yx, yX, Yx, YX) being declared 'transpose', (Xy) was.
As a result, we would detect symmetries that were not present as well
as failing to detect those that were whenever the x- and y-axis span
were not the same.

Happily, fixing that shows a net reduction in the number of iterations,
so we can run about 10% faster.