A051602: add ./classify
Pipe output of `./try -v <n>` or `./try -m <n> to this program to check
for duplicates (under all symmetries, including order-4 symmetries).

Without options, reports pairs (by reported row number) that are the same.
With '-u', pretty-prints the unique arrangements; with '-u -i', also
indexes the elements in the pretty-print (0-9, A-Z, a-z).