A051602: consistent symmetry breaking
Some aspects of our deduplication are based on ordering points based
on the order they were added to the arrangement; other aspects seek
to distinguish not-yet-added points, and order them based on their
Cartesian coordinates.

In some cases the two approaches can trip over each other, such that
we can end up trying both of two equivalent arrangements (slow) or
trying neither of them (disastrous).

So we replace the coordinate-based code with inclusion-order. We do
this by a) moving the relevant functions (sym_best, sym_best2) from
sym.c to try.c where they have access to the point[] array, and b)
replacing sym_best (which asked if this 1-point extension being
considered is canonically "best") with sym_best3 (which asks if the
3 points providing that 1-point extension are canonically best).