A051602: track seen pairs
Replace c6027543b7 ("avoid more duplicates"), which avoided trying the
same pair twice at the same level, which a more generic tracking of
pairs by means of new types pair_t and pairlist_t.

This additionally allows us a) to propagate the list of seen pairs
through to later contexts, and b) allows us to promote a seen pair
to a seen single point when the other point of the pair is seen.