A051602: rot45 in classify where possible
The expectation is that arrangements emitted by ./try will have all
points even, so we divide them all by 2. We do not expect the results
to be all even, but they may be all same-parity (ie x- and y-coordinate
having the same value (mod 2)). In that case we spot more duplicates by
rotating the arrangement 45 degrees and reducing by a factor of sqrt(2).