A051602: include 2-point extensions for diagonals
For each pair of points we tried treating it as the two points on the
left side of a square, and those on the right side; we failed to
include treating it as two points diagonally opposed in a square.

Rectify that omission, taking advantage of the power-of-2 scaling
already added to ensure all points and potential points can continue
to be treated as having integer coordinates.