divrep harness: safer strategy calculation
Within a run, we can go from a specified minimum (optn) to a specified
maximum (optx); but if we stop early (eg because we found a solution)
we do not change the entry for the run. To allow for this, we instead
have used taug's "checked (to)" value to tell us not only where we
should start searching from in the next run but also to determine
the range actually checked by the previous run (so as to infer timings
used to determine what range to target in the next run).

This is no longer safe if some other process may have done work that
directly results in an update to taug.checked.

A more robust fix would probably be to update the run entry after all,
either optx or a new field; for now however we fall back to assuming
the original optx whenever we find taug.checked is greater.