divrep oul: first pass
For oneseq (-yo), aim to find f(n,k) by directly generating all possible
candidates. This first pass fully handles f(n,1), and handles f(n,2)
when n is of the form 2p.

This generates Seq::Run objects with an owner of a new type, 'upperlim';
the intention is that for harder cases this will progressively find
and record improved upper bounds for general f(n,k).

It is possible that some aspects of this can also later be generalised
to other sequence types.