divrep oul: deduplicate primes for repeated factors
Replace index i with array of primes p[]; if this factor is the same
as the previous one, skip primes <= p[-1]. The reduced duplication
improves speed calculating f(n,2) for n in (2, 4, ... 100) from 724s
to 600s.

Additionally replace fi (index in $fa) with f (value from $fa).

We still see duplicates for multiples of 9, which are not a problem:
they occur because we have a hack to avoid setting factors of 3, but
do not elide those factors in the list of factorizations. Replacing
the hack as described in bfe9d5e9ab is expected to fix those.