divrep oul: generalize order of selection of factorizations
Given we want all 2-tuples of integers in (0 .. size-1), we a) prefer
to minimize sum of tuple values, and lexicographically order only within
a given sum, and b) take advantage of the opportunity to generalize for
n-tuples.

Hopefully this will easy to adapt if we later decide (also, or instead)
to minimize the maximum tuple value - so that (1, 1) might come before
(0, 2), and optionally (1, 1, 1) might come before (0, 0, 2).