divrep oul: bugfix, check q.taumin > best
Given q the total allocated so far, p^x the new allocation, and taumin
the least the would need to be allocated to satisfy the remaining tau,
the intent was to check whether q.p^x.taumin was already greater than
the best result found so far. In error, we were checking p^x.taumin
instead.

We now include p^x in q earlier, and check for q.taumin > best.