divrep oul: strategize walk over allocation
When about to embark on a new allocation, given lim the greatest value
we may have to allocate (calculated as (best / q / mintau)^(1/x)) and
walk the number of values we'd walk if we allocated nothing more,
choose the walk whenever walk < lim.

This finally replaces the "don't extend a factor of 3" hack; with all the
recent changes, total time to calculate f(n,2) for n in (2, 4, ... 100)
reduces from 600s to 46s.