divrep report: show timing for full run; add -r option
If we have a completed run of type upperlim, show its timing on the
report_f() lines.

Additionally, by default treat 'report <n> <k>' as a request to show
just the one line of results for f(n,k). To get the previous behaviour
showing details of all runs for that value, supply the new '-r' option.