divrep oul: wrongly avoiding primes can underestimate mintau
Bugfix: if calculating mintau() for v_i while avoiding prime p because
it is allocated for v_{i+p}, we will wrongly reject a prime that _must_
divide v_i.

Code now checking for this is ugly, there must be a better way; however
the need for it may well disappear when we add preallocation of small
primes.