divrep oul: explicitly avoid bad multiples of 4
Specifically check for combinations of i and x such that 2^x cannot
divide v_i because that would give v_j == 4 (mod 8) for some j, and
n does not divide 3. Use this to set the first prime to try for such
(i, x) combinations to 3 instead of 2.

We only need to consider p=2 for this, since max(k) < 3^2 whenever
n does not divide 3.