divrep: store partial results
Mark runs as "partial" if they are not intended as proof that a result
found is minimal, and record such results as an upper bound on f(n,k).

For now we leave taug.ming to represent the highest k for which the
minimal value has been established, this still needs to be fixed.
./upperlim is also unchanged, but could usefully now record any result
it finds.

./report will show eg "f(84, 4) <= 12965471279446575671872";
./oul will record partial results if run with -p; Seq::TauF::good
(and the new ...::partial) will update tauf.f only if it is an
improvement.