divrep oul: in walk_v, order factorizations smallest first
We already categorize the factorizations to put those that have faster
special-case code first: squares, then primes, then semiprimes, then
the rest. Within those categories we now also sort them to put the
smallest values first, on the assumption that we can factorize them
(and thus have the opportunity to short-circuit the rest of the list
if they don't match) the fastest.