divrep oul: force related powers
At the point we allocate p^x to v_i, force appropriate powers of p for
every other v_j with i == j (mod p).

Where a specific power is forced, we allocate that immediately; where we
know only the minimum power that should be forced, we instead mark the
fact in the structure for v_j, and consider only when best_v() decides
something should be allocated here.