divrep: bugfix, gtauseq can wrongly fix_power
In generalizing the support for tauseq to addseq and oneseq, one of
the tauseq-specific aspects was missed: if we know, say, that some
element v_i = 0 (mod 3) that tells us that we have a "floating" factor
of 3: the 3-adic order o_3(v_i) >= 1; if we know that v_i = 3 (mod 9),
however, we have a "fixed" factor of 3: o_3(v_i) = 1. Fixed and floating
powers are used primarily to detect an opportunity for the "fixed power"
optimization.

For tauseq, we additionally know that v_i = n + di, so when we have
v_i = 0 (mod 3), n == 0 (mod 3^2), d != 0 (mod 3), we additionally say
that o_3(v_i) = 1 (not >= 1): we account separately in float_spare()
for the possibility that i == 0 (mod 3). This is not the case for
oneseq and addseq, but the logic was shared across all three.

This has been fixed by making Type::Tauish::is_fixed() a method,
moving the wrongly shared version to Type::TauSeq, and leaving a
simpler shared version in its place.

Luckily this hasn't affected any actual runs: it was noticed for
the case of oneseq(18,5), which was beyond the range we had attempted
with gtauseq, when attempting to confirm aspects of the run with oul.