divrep oul: add option -f (force full)
By default for "forced" primes we try setting them to each qualifying
power in turn that divides the outstanding tau _and_ satisfies at least
one of the odd prime divisors of tau. When we're down to satisfying only
powers of 2, we effectively give up and fall through to the normal prime
allocation code (knowing that it won't attempt an additional allocation
of the "forced" prime, since it is marked as used).

If run with the new -f ('force full') option, we will instead try
allocating even p^{2^x - 1}, removing the need to fall through.